More faithfully implement Chaikin's method
Originally this algorithm attempted to maintain the first and last
points when iterating through Chaikin's method. However, in the real
algorithm, the first point slowly edges away from its original position.

If we start with n+1 points, each iteration should give us 2n points[1].
However, we were getting 2n+2 points.

This problem compounds with each step of the iteration, so we were
gaining many points at the beginning and end of the sequence. The
additional points at the beginning and end were all along a straight
line with each other and therefore didn't do much to improve the final
appearance of the curve. Calling code may want to restore the first and
last points to maintain the appearance of the original implementation.

[1] http://graphics.cs.ucdavis.edu/education/CAGDNotes/Chaikins-Algorithm/Chaikins-Algorithm.html